Fixed errors, 4gb working in 30 sec

+++ task2/hw2Proshletsov/lib.h
@@ -30,7 +30,7 @@ typedef struct Error_hndler
 
 typedef struct Op_table
 {
-    size_t (*recieve) (Pipe *self, int max_size);
+    size_t (*recieve) (Pipe *self, size_t max_size);
     size_t    (*send) (Pipe *self, size_t amount_to_send, pid_t pid);
 } Ops;
 
@@ -47,11 +47,11 @@ typedef struct P_pipe
 
 Pipe *pipe_construct (size_t size);
 
-size_t recieve_op (Pipe *self, int max_size);
+size_t recieve_op (Pipe *self, size_t max_size);
 size_t    send_op (Pipe *self, size_t amount_to_send, pid_t pid);
 
 void tune_pipes_for_child  (Pipe *pipe);
 void tune_pipes_for_parent (Pipe *pipe);
 
-int send_file_size (FILE* file, Pipe *pipe);
-int recieve_file_size (Pipe *pipe);
+size_t send_file_size (FILE* file, Pipe *pipe);
+size_t recieve_file_size (Pipe *pipe);
